+++ R/get_datagrid.R
@@ -42,13 +42,18 @@
 #'   - a "token" that creates pre-defined representative values:
 #'     - for mean and -/+ 1 SD around the mean: `"x = [sd]"`
 #'     - for median and -/+ 1 MAD around the median: `"x = [mad]"`
-#'     - for Tukey's five number summary (minimum, lower-hinge, median, upper-hinge, maximum): `"x = [fivenum]"`
+#'     - for Tukey's five number summary (minimum, lower-hinge, median,
+#'       upper-hinge, maximum): `"x = [fivenum]"`
 #'     - for terciles, including minimum and maximum: `"x = [terciles]"`
 #'     - for terciles, excluding minimum and maximum: `"x = [terciles2]"`
-#'     - for quartiles, including minimum and maximum: `"x = [quartiles]"`
+#'     - for quartiles, including minimum and maximum: `"x = [quartiles]"` (same
+#'       as `"x = [fivenum]"`)
 #'     - for quartiles, excluding minimum and maximum: `"x = [quartiles2]"`
+#'     - for a pretty value range: `"x = [pretty]"`
 #'     - for minimum and maximum value: `"x = [minmax]"`
 #'     - for 0 and the maximum value: `"x = [zeromax]"`
+#'     - for a random sample from all values: `"x = [sample <number>]"`, where
+#'       `<number>` should be a positive integer, e.g. `"x = [sample 15]"`.
 #'
 #'   For **factor** variables, the value(s) inside the brackets should indicate
 #'   one or more factor levels, like `by = "Species = [setosa, versicolor]"`.
@@ -261,6 +266,12 @@ get_datagrid.data.frame <- function(x,
     specs$varname <- as.character(specs$varname) # make sure it's a string not fac
     specs <- specs[!duplicated(specs$varname), ] # Drop duplicates
 
+    # sanity check - target in data?
+    if (!all(specs$varname %in% colnames(x))) {
+      format_error(paste0(
+        "Variable `", setdiff(specs$varname, colnames(x))[1], "` was not found in the data. Please check the spelling." # nolint
+      ))
+    }
     specs$is_factor <- vapply(x[specs$varname], function(x) is.factor(x) || is.character(x), TRUE)
 
     # Create target list of factors -----------------------------------------
@@ -517,7 +528,7 @@ get_datagrid.default <- function(x,
   by <- .extract_at_interactions(by)
 
   # Drop random factors
-  random_factors <- find_random(x, flatten = TRUE)
+  random_factors <- find_random(x, flatten = TRUE, split_nested = TRUE)
   if (!is.null(random_factors)) {
     if (isFALSE(include_random)) {
       # drop random factors, if these should not be included
@@ -635,7 +646,6 @@ get_datagrid.visualisation_matrix <- function(x, reference = attributes(x)$refer
   datagrid
 }
 
-
 #' @export
 get_datagrid.datagrid <- get_datagrid.visualisation_matrix
 
@@ -699,6 +709,7 @@ get_datagrid.emmGrid <- function(x, ...) {
   data.frame(s)[, which_cols, drop = FALSE]
 }
 
+
 #' @export
 get_datagrid.emm_list <- function(x, ...) {
   k <- length(x)
@@ -716,6 +727,7 @@ get_datagrid.emm_list <- function(x, ...) {
   clear_cols <- colnames(out)[sapply(out, Negate(anyNA))] # these should be first
   out[, c(clear_cols, setdiff(colnames(out), clear_cols)), drop = FALSE]
 }
+
 
 #' @export
 get_datagrid.slopes <- function(x, ...) {
@@ -780,20 +792,26 @@ get_datagrid.comparisons <- get_datagrid.slopes
       if (all(grepl('\\".*\\"', parts))) parts <- gsub('"', "", parts, fixed = TRUE)
 
       # Make expression ----------
-      if (is.factor(x) || is.character(x)) {
+      shortcuts <- c(
+        "meansd", "sd", "mad", "quartiles", "quartiles2", "zeromax",
+        "minmax", "terciles", "terciles2", "fivenum", "pretty"
+      )
+      if ((is.factor(x) && all(parts %in% levels(x))) || (is.character(x) && all(parts %in% x)))   {
         # Factor
         # Add quotes around them
         parts <- paste0("'", parts, "'")
         # Convert to character
         by_expression <- paste0("as.factor(c(", toString(parts), "))")
       } else if (length(parts) == 1) {
-        # Numeric
-        # If one, might be a shortcut
-        shortcuts <- c(
-          "meansd", "sd", "mad", "quartiles", "quartiles2", "zeromax",
-          "minmax", "terciles", "terciles2", "fivenum"
-        )
-        if (parts %in% shortcuts) {
+        # If one, might be a shortcut. or a sampling request
+        if (grepl("sample", parts, fixed = TRUE)) {
+          n_to_sample <- .safe(as.numeric(trim_ws(gsub("sample", "", parts, fixed = TRUE))))
+          # do we have a proper definition of the sample size? If not, error
+          if (is.null(n_to_sample) || is.na(n_to_sample) || !length(n_to_sample)) {
+            format_error("The token `sample` must be followed by the number of samples to be drawn, e.g. `[sample 15]`.") # nolint
+          }
+          by_expression <- paste0("c(", paste(sample(x, n_to_sample), collapse = ","), ")")
+        } else if (parts %in% shortcuts) {
           if (parts %in% c("meansd", "sd")) {
             center <- mean(x, na.rm = TRUE)
             spread <- stats::sd(x, na.rm = TRUE)
@@ -802,16 +820,16 @@ get_datagrid.comparisons <- get_datagrid.slopes
             center <- stats::median(x, na.rm = TRUE)
             spread <- stats::mad(x, na.rm = TRUE)
             by_expression <- paste0("c(", center - spread, ",", center, ",", center + spread, ")")
-          } else if (parts == "quartiles") {
-            by_expression <- paste0("c(", paste(as.vector(stats::quantile(x, na.rm = TRUE)), collapse = ","), ")")
-          } else if (parts == "quartiles2") {
-            by_expression <- paste0("c(", paste(as.vector(stats::quantile(x, na.rm = TRUE))[2:4], collapse = ","), ")")
-          } else if (parts == "terciles") {
-            by_expression <- paste0("c(", paste(as.vector(stats::quantile(x, probs = (1:2) / 3, na.rm = TRUE)), collapse = ","), ")") # nolint
-          } else if (parts == "terciles2") {
-            by_expression <- paste0("c(", paste(as.vector(stats::quantile(x, probs = (0:3) / 3, na.rm = TRUE)), collapse = ","), ")") # nolint
-          } else if (parts == "fivenum") {
+          } else if (parts %in% c("fivenum", "quartiles")) {
             by_expression <- paste0("c(", paste(as.vector(stats::fivenum(x, na.rm = TRUE)), collapse = ","), ")")
+          } else if (parts == "quartiles2") {
+            by_expression <- paste0("c(", paste(as.vector(stats::quantile(x, na.rm = TRUE))[2:4], collapse = ","), ")") # nolint
+          } else if (parts == "terciles") {
+            by_expression <- paste0("c(", paste(as.vector(stats::quantile(x, probs = (0:3) / 3, na.rm = TRUE)), collapse = ","), ")") # nolint
+          } else if (parts == "terciles2") {
+            by_expression <- paste0("c(", paste(as.vector(stats::quantile(x, probs = (1:2) / 3, na.rm = TRUE)), collapse = ","), ")") # nolint
+          } else if (parts == "pretty") {
+            by_expression <- paste0("c(", paste(as.vector(pretty(x, na.rm = TRUE)), collapse = ","), ")")
           } else if (parts == "zeromax") {
             by_expression <- paste0("c(0,", max(x, na.rm = TRUE), ")")
           } else if (parts == "minmax") {
@@ -820,21 +838,28 @@ get_datagrid.comparisons <- get_datagrid.slopes
         } else if (is.numeric(parts)) {
           by_expression <- parts
         } else {
-          format_error(
-            paste0(
-              "The `by` argument (", by, ") should either indicate the minimum and the maximum, or one of the following options: ", # nolint
-              toString(shortcuts),
-              "."
-            )
-          )
+          by_expression <- NULL
         }
         # If only two, it's probably the range
-      } else if (length(parts) == 2) {
-        by_expression <- paste0("seq(", parts[1], ", ", parts[2], ", length.out = length)")
-        # If more, it's probably the vector
-      } else if (length(parts) > 2L) {
-        parts <- as.numeric(parts)
-        by_expression <- paste0("c(", toString(parts), ")")
+      } else if (is.numeric(x)) {
+        if (length(parts) == 2) {
+          by_expression <- paste0("seq(", parts[1], ", ", parts[2], ", length.out = length)")
+          # If more, it's probably the vector
+        } else if (length(parts) > 2L) {
+          parts <- as.numeric(parts)
+          by_expression <- paste0("c(", toString(parts), ")")
+        }
+      } else {
+        by_expression <- NULL
+      }
+      if (is.null(by_expression)) {
+        format_error(
+          paste0(
+            "The `by` argument (", by, ") should either indicate a valid factor level, the minimum and the maximum value of a vector, or one of the following options: ", # nolint
+            toString(shortcuts),
+            "."
+          )
+        )
       }
       # Else, try to directly eval the content
     } else {
@@ -857,6 +882,7 @@ get_datagrid.comparisons <- get_datagrid.slopes
   }
   data.frame(varname = varname, expression = by_expression, stringsAsFactors = FALSE)
 }
+
 
 #' @keywords internal
 .get_datagrid_summary <- function(x, numerics = "mean", factors = "reference", na.rm = TRUE, ...) {
@@ -901,6 +927,7 @@ get_datagrid.comparisons <- get_datagrid.slopes
   }
   out
 }
+
 
 #' @keywords internal
 .create_spread <- function(x, length = 10, range = "range", ci = 0.95, ...) {
@@ -965,6 +992,7 @@ get_datagrid.comparisons <- get_datagrid.slopes
   seq(mini, maxi, length.out = length)
 }
 
+
 #' @keywords internal
 .data_match <- function(x, to, ...) {
   if (!is.data.frame(to)) {
@@ -978,6 +1006,7 @@ get_datagrid.comparisons <- get_datagrid.slopes
   }
   .to_numeric(row.names(x)[idx])
 }
+
 
 #' @keywords internal
 .get_model_data_for_grid <- function(x, data) {